refactor: update icon imports to use icons.constants and improve file handling in FileConverter
