+++ src/app/(protected)/audio/page.tsx
@@ -14,7 +14,7 @@ export default function FileConverter() {
                 <h1 className="font-alegreya text-text-primary text-4xl tracking-wide">File Converter</h1>
                 <p className="text-highlight mt-2 mb-10 text-lg">Convert your audio files to any format</p>
 
-                {files?.length === 0 && <FileUpload onFilesSelected={setFiles} maxFiles={10} accept="audio/*" />}
+                {(files ?? []).length === 0 && <FileUpload onFilesSelected={setFiles} maxFiles={10} accept="audio/*" />}
 
                 {files?.map((file, index) => (
                     <FileUploadItem key={index} file={file} onRemove={() => setFiles((prev) => (prev ? prev.filter((f) => f !== file) : null))} />
