+++ src/constants/icons.constants.ts
@@ -4,11 +4,13 @@ enum IconSet {
     manga = 'custom:manga',
     musicConvert = 'custom:musicConvert',
 
+    '3d' = 'file-icons:3d-model',
     /* ------------------------------------ A ----------------------------------- */
     anilist = 'simple-icons:anilist',
     arrowOpenLeft = 'line-md:arrow-open-left',
     arrowOpenRight = 'line-md:arrow-open-right',
-    audio = 'solar:music-note-bold',
+    audio = 'solar:headphones-round-bold-duotone',
+    archive = 'solar:zip-file-linear',
 
     /* ------------------------------------ B ----------------------------------- */
     broom = 'game-icons:broom',
@@ -18,9 +20,13 @@ enum IconSet {
     circle = 'ci:circle',
     close = 'iconamoon:close',
     closeAnimated = 'line-md:close-small',
+    code = 'solar:code-file-linear',
 
     /* ------------------------------------ D ----------------------------------- */
+    database = 'solar:database-linear',
+    'disk image' = 'solar:diskette-linear',
     desktop = 'noto:desktop-computer',
+    document = 'solar:document-text-linear',
     down = 'formkit:down',
     draw = 'fa-solid:handshake',
 
@@ -29,12 +35,14 @@ enum IconSet {
     editProfile = 'fluent:person-edit-24-filled',
     email = 'line-md:email-alt-filled',
     error = 'meteocons:code-red-fill',
+    executable = 'carbon:executable-program',
     eye = 'line-md:watch-loop',
     eyeClose = 'line-md:watch-off-loop',
 
     /* ------------------------------------ F ----------------------------------- */
     filePlus = 'line-md:file-plus-filled',
     filter = 'line-md:filter-filled',
+    font = 'hugeicons:tiff-02',
 
     /* ------------------------------------ G ----------------------------------- */
     gamepad = 'emojione-monotone:video-game',
@@ -46,6 +54,7 @@ enum IconSet {
     heart = 'icomoon-free:heart',
 
     /* ------------------------------------ I ----------------------------------- */
+    image = 'line-md:image-filled',
     info = 'ic:round-info',
     infoOutlined = 'carbon:information',
 
@@ -60,6 +69,7 @@ enum IconSet {
     logout = 'majesticons:logout',
 
     /* ------------------------------------ M ----------------------------------- */
+    markup = 'garden:markup-stroke-12',
     menu = 'hugeicons:menu-02',
     minus = 'line-md:minus',
     moon = 'meteocons:clear-night-fill',
@@ -74,6 +84,7 @@ enum IconSet {
     personAdd = 'line-md:person-add-filled',
     player = 'game-icons:centurion-helmet',
     plus = 'line-md:plus',
+    presentation = 'hugeicons:presentation-bar-chart-02',
 
     /* ------------------------------------ R ----------------------------------- */
     right = 'formkit:right',
@@ -81,22 +92,29 @@ enum IconSet {
     refresh = 'grommet-icons:power-reset',
 
     /* ------------------------------------ S ----------------------------------- */
+    script = 'proicons:script',
     search = 'line-md:search',
     settings = 'solar:settings-bold',
     settingsOutlined = 'solar:settings-outline',
     share = 'solar:share-bold-duotone',
     smile = 'line-md:emoji-smile',
     spotify = 'line-md:spotify-filled',
+    spreadsheet = 'tabler:file-spreadsheet',
     success = 'line-md:confirm-circle-filled',
     sun = 'meteocons:clear-day-fill',
 
     /* ------------------------------------ T ----------------------------------- */
+    text = 'solar:file-text-linear',
     trash = 'entypo:trash',
 
     /* ------------------------------------ U ----------------------------------- */
     up = 'formkit:up',
     unlink = 'octicon:unlink-16',
     upload = 'line-md:cloud-alt-upload-filled-loop',
+
+    /* ------------------------------------ V ----------------------------------- */
+    video = 'tabler:video-filled',
+    vector = 'bi:filetype-psd',
 
     /* ------------------------------------ W ----------------------------------- */
     warning = 'meteocons:code-yellow-fill',
